+++ examples/cf-worker/src/index.ts
@@ -11,7 +11,7 @@
  * Learn more at https://developers.cloudflare.com/workers/
  */
 import * as logfire from '@pydantic/logfire-api';
-import { instrumentTail } from '@pydantic/logfire-cf-workers';
+import { instrument } from '@pydantic/logfire-cf-workers';
 
 const handler = {
 	async fetch(): Promise<Response> {
@@ -20,7 +20,7 @@ const handler = {
 	},
 } satisfies ExportedHandler;
 
-export default instrumentTail(handler, {
+export default instrument(handler, {
 	service: {
 		name: 'cloudflare-worker',
 		namespace: '',