+++ examples/node/index.ts
@@ -5,16 +5,29 @@ logfire.configure({
   serviceName: 'example-node-script',
   serviceVersion: '1.0.0',
   environment: 'staging',
-  diagLogLevel: logfire.DiagLogLevel.INFO,
-  codeSource: {
-    repository: 'https://github.com/pydantic/pydantic',
-    revision: 'master',
-  },
+  diagLogLevel: logfire.DiagLogLevel.DEBUG,
 })
 
 
-logfire.info('Hello from Node.js', {
-  'attribute-key': 'attribute-value'
+logfire.span('Hello from Node.js, {next_player}', {
+  'attribute-key': 'attribute-value',
+  next_player: '0',
+  arr: [1, 2, 3],
+  something: {
+    value: [1, 2, 3],
+    key: 'value'
+  }
 }, {
   tags: ['example', 'example2']
+}, (span) => {
+  span.end()
 })
+
+if (process.env.TRIGGER_ERROR) {
+  try {
+    throw new Error('This is an error for testing purposes');
+  } catch (error) {
+    logfire.reportError("An error occurred", error as Error);
+    console.error("An error occurred:", error);
+  }
+}