+++ packages/logfire-api/src/index.ts
@@ -5,11 +5,12 @@ import { ATTR_EXCEPTION_MESSAGE, ATTR_EXCEPTION_STACKTRACE } from '@opentelemetr
 import { ScrubCallback } from './AttributeScrubber'
 import { ATTRIBUTES_LEVEL_KEY, ATTRIBUTES_MESSAGE_TEMPLATE_KEY, ATTRIBUTES_SPAN_TYPE_KEY, ATTRIBUTES_TAGS_KEY } from './constants'
 import { logfireFormatWithExtras } from './formatter'
-import { logfireApiConfig } from './logfireApiConfig'
+import { logfireApiConfig, serializeAttributes } from './logfireApiConfig'
 
 export * from './AttributeScrubber'
 export { configureLogfireApi, logfireApiConfig, resolveBaseUrl, resolveSendToLogfire } from './logfireApiConfig'
 export { serializeAttributes } from './serializeAttributes'
+export * from './ULIDGenerator'
 
 export interface SrubbingOptions {
   callback?: ScrubCallback
@@ -53,7 +54,7 @@ export function startSpan(
     formattedMessage,
     {
       attributes: {
-        ...attributes,
+        ...serializeAttributes(attributes),
         [ATTRIBUTES_MESSAGE_TEMPLATE_KEY]: newTemplate,
         [ATTRIBUTES_LEVEL_KEY]: level,
         [ATTRIBUTES_TAGS_KEY]: Array.from(new Set(tags).values()),
@@ -67,16 +68,18 @@ export function startSpan(
 }
 
 export function span<F extends (span: Span) => unknown>(
-  message: string,
+  msgTemplate: string,
   attributes: Record<string, unknown> = {},
   { tags = [], level = Level.Info }: LogOptions = {},
   callback: F
 ) {
+  const [formattedMessage, , newTemplate] = logfireFormatWithExtras(msgTemplate, attributes, logfireApiConfig.scrubber)
   return logfireApiConfig.tracer.startActiveSpan<F>(
-    message,
+    formattedMessage,
     {
       attributes: {
-        ...attributes,
+        ...serializeAttributes(attributes),
+        [ATTRIBUTES_MESSAGE_TEMPLATE_KEY]: newTemplate,
         [ATTRIBUTES_LEVEL_KEY]: level,
         [ATTRIBUTES_TAGS_KEY]: Array.from(new Set(tags).values()),
       },
@@ -117,13 +120,23 @@ export function warning(message: string, attributes: Record<string, unknown> = {
   log(message, attributes, { ...options, level: Level.Warning })
 }
 
+/**
+ * Use this method to report an error to Logfire.
+ * Captures the error stack trace and message in the respective semantic attributes and sets the correct level and status.
+ */
 export function reportError(message: string, error: Error, extraAttributes: Record<string, unknown> = {}) {
-  const span = startSpan(message, {
-    // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-    [ATTR_EXCEPTION_MESSAGE]: error.message ?? 'error',
-    [ATTR_EXCEPTION_STACKTRACE]: error.stack,
-    ...extraAttributes,
-  })
+  const span = startSpan(
+    message,
+    {
+      // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+      [ATTR_EXCEPTION_MESSAGE]: error.message ?? 'error',
+      [ATTR_EXCEPTION_STACKTRACE]: error.stack,
+      ...extraAttributes,
+    },
+    {
+      level: Level.Error,
+    }
+  )
 
   span.recordException(error)
   span.setStatus({ code: SpanStatusCode.ERROR })