+++ packages/logfire-cf-workers/src/index.ts
@@ -1,10 +1,9 @@
 import type { ReadableSpan } from '@opentelemetry/sdk-trace-base'
 
-import { resolveBaseUrl, serializeAttributes } from '@pydantic/logfire-api'
+import { resolveBaseUrl, serializeAttributes, ULIDGenerator } from '@pydantic/logfire-api'
 import { instrument as baseInstrument, TraceConfig } from '@pydantic/otel-cf-workers'
 
 import { TailWorkerExporter } from './TailWorkerExporter'
-import { ULIDGenerator } from './ULIDGenerator'
 export * from './exportTailEventsToLogfire'
 
 export interface CloudflareConfigOptions {
@@ -14,7 +13,10 @@ export interface CloudflareConfigOptions {
 
 type Env = Record<string, string | undefined>
 
-type ConfigOptionsBase = Pick<TraceConfig, 'fetch' | 'handlers' | 'instrumentation' | 'propagator' | 'sampling' | 'scope' | 'service'>
+type ConfigOptionsBase = Pick<
+  TraceConfig,
+  'environment' | 'fetch' | 'handlers' | 'instrumentation' | 'propagator' | 'sampling' | 'scope' | 'service'
+>
 
 export interface InProcessConfigOptions extends ConfigOptionsBase {
   baseUrl?: string
@@ -25,11 +27,13 @@ export interface TailConfigOptions extends ConfigOptionsBase {}
 
 function getInProcessConfig(config: InProcessConfigOptions): (env: Env) => TraceConfig {
   return (env: Env): TraceConfig => {
-    const { LOGFIRE_TOKEN: token = '' } = env
+    const { LOGFIRE_ENVIRONMENT: envDeploymentEnvironment, LOGFIRE_TOKEN: token = '' } = env
 
     const baseUrl = resolveBaseUrl(env, config.baseUrl, token)
+    const resolvedEnvironment = config.environment ?? envDeploymentEnvironment
 
     return Object.assign({}, config, {
+      environment: resolvedEnvironment,
       exporter: {
         headers: { Authorization: token },
         url: `${baseUrl}/v1/traces`,
@@ -57,6 +61,11 @@ export function instrumentInProcess<T>(handler: T, config: InProcessConfigOption
 export function instrumentTail<T>(handler: T, config: TailConfigOptions): T {
   return baseInstrument(handler, getTailConfig(config)) as T
 }
+
+/**
+ * Alias for `instrumentInProcess` to maintain compatibility with previous versions.
+ */
+export const instrument = instrumentInProcess
 
 function postProcessAttributes(spans: ReadableSpan[]) {
   for (const span of spans) {